Triton/requirements.txt
tritonclient[all]
transformers
pillow
torch